added rebalance_gpu.cu
