+++ domain/include/cstone/focus/CMakeLists.txt
@@ -1,8 +1,8 @@
 if(CMAKE_HIP_COMPILER)
-    set_source_files_properties(source_center_gpu.cu PROPERTIES LANGUAGE HIP)
+    set_source_files_properties(rebalance_gpu.cu source_center_gpu.cu PROPERTIES LANGUAGE HIP)
 endif()
 
 if(CMAKE_CUDA_COMPILER OR CMAKE_HIP_COMPILER)
-    add_library(source_center_gpu_obj OBJECT source_center_gpu.cu)
+    add_library(source_center_gpu_obj OBJECT rebalance_gpu.cu source_center_gpu.cu)
     target_include_directories(source_center_gpu_obj PRIVATE ${PROJECT_SOURCE_DIR}/include)
 endif()
